requirements.txt
certifi==2018.1.18
chardet==3.0.4
future==0.16.0
idna==2.6
PyExecJS==1.5.1
python-telegram-bot==10.0.1
requests==2.18.4
six==1.11.0
urllib3==1.22
